Move queries to different file and add descriptions to functions
